refactor Signup into PasswordSignupConnected to follow connected component pattern
