+++ packages/apollos-ui-auth/src/PasswordSignup/PasswordSignup.tests.js
@@ -7,7 +7,9 @@ import { ApolloProvider } from 'react-apollo';
 import { ApolloClient } from 'apollo-client';
 import { InMemoryCache } from 'apollo-cache-inmemory';
 import { createHttpLink } from 'apollo-link-http';
-import { resolvers } from '../../Provider';
+
+import { resolvers } from '../Provider';
+
 import PasswordSignup from './PasswordSignup';
 
 const cache = new InMemoryCache();
