+++ packages/apollos-ui-auth/src/PasswordSignup/index.js
@@ -1,0 +1,4 @@
+import PasswordSignup from './PasswordSignup';
+
+export PasswordSignupConnected from './PasswordSignupConnected';
+export default PasswordSignup;
